content: auto-complete from the search dialog
Auto-completion is added to the search dialog, both on the home page and
in the navbar. This is achieved using autoComplete.js for autocompletion
behavior, but with heavy modification to make the component accessible.

Additionally, some DOM changes are made to support the visual styling of
the auto-completion dialog:
 + The SearchForm and Header-searchForm components are made
   position:absolute.
 + Relatively positioned container components are added to control the
   positioning and spacing of the search forms.

The header search form is made to have a white background, as the blue
background looked incorrect when the completion list overlaid the white
of the body content.

Fixes b/143370178

Change-Id: Ic0b527643c8c5aa6e04101082e7e630325486802
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/604267
Reviewed-by: Alexander Rakoczy <rakoczy@google.com>