internal/etl: add an action to populate completion indexes
We're going to compute auto-completions using redis sorted sets in the
'HA' redis instance (because we don't want them to be evicted by
caching). This data need not be very fresh, and so can be computed by
the ETL every so often, perhaps once a day.

This CL wires the HA redis instance into the ETL, and adds an action to
update these Redis sorted sets. Doing this takes 1-2 minutes and
consumes quite a lot of memory, so there are some optimizations used:
 + the underlying *sql.DB is used for streaming search documents, rather
   than a method on postgres.DB. There is explanation for this in the
   code, and additionally I don't think this is so bad as long as it is
   only done in the ETL.
 + Completions data is loaded into temporary Redis keys and then renamed
   to overwrite the keys that will be used by the frontend, in order to
   minimize disruption while the completion data is being updated.

Updates b/143370178

Change-Id: I48caf77eb26d8ea7b60c2e117481bfa98ef089ca
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/602062
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>