internal/frontend: add an autocompletion endpoint
Now that completion data is being populated by the ETL, we can use it to
provide a JSON API for search autocompletions at /autocomplete?q=query.

This API reads lexical completions from Redis, sorts according to a
scoring heuristic, and returns the results as a JSON array.

For testing purposes, the pathCompletions function is moved from the etl
package to the complete package, and exported.

Updates b/143370178

Change-Id: I61d75296399222fa9ebfe8aadc909bf048389a23
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/602064
Reviewed-by: Jonathan Amsterdam <jba@google.com>