internal/frontend: make latest-version links canonical
Put the link in the latest-version badge in the form
mod@version/suffix.

The best way to do this was to move the latest-badge-link
logic from the template to the server itself.

Also, rename the Suffix field of Package to PathAfterDirectory
to more accurately describe it.

Fixes b/144681389.

Change-Id: I16a78224ee6fa53b60954716338347a298dbef21
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/604414
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>