internal/etl/dochtml: construct anchor points more carefully
Each exported symbol in the package has an anchor point: an HTML "id"
attribute that can used as a URL fragment, in order to link or jump to
that symbol.

Previously, we were generating two identical "id" attributes for some
symbols; this CL fixes that.

It also adds a "data-kind" attribute on the same elements for which we
provide an "id", so that tools that offer a list of symbols can
display each symbol's kind (constant, variable, etc.) as well.

I originally intended to make this two separate CLs, but it turns out
the duplicate-ID bug and data-kind feature are closely related, since
we want one if and only if we have the other.

Fixes b/143557716.
Fixes b/143456714.

Change-Id: I680e126d3720871a980664a06d7c60ad414c4466
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/605504
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>