internal/version: construct sortable version strings
Add version.ForSorting, which constructs an ASCII string from
a version that has the correct semver ordering.

With this, we can replace our four semver columns with one,
and use a "<" operator instead of a more complicated expression.

This encoding also fixes a small bug in the previous one (see
b/139482696), and removes limits like the 20-character maximum for
padding a prerelease number, and the requirement that the major, minor
and patch versions fit into machine integers.

Sometimes we don't follow semver ordering, because we want
to list release versions before prereleases. We've been doing
that with an expression like

  CASE WHEN prerelease = '~' THEN 0 ELSE 1 END

in the ORDER BY clause. We can instead use

 CASE WHEN right(sort_version, 1) = '~' THEN 0 ELSE 1 END

because only release versions end in '~'.

Updates b/139482696.

Change-Id: If24a95cec8ec196ae22fff72fc2bd7fecd79028c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/605517
Reviewed-by: Robert Findley <rfindley@google.com>