internal/frontend: distinguish titles
Distinguish the "page title," which is put in the <title> element
and thus appears in tabs, from the title that appears at the top
of a details page.

The title in a tab should start with the resource name (e.g. "io package")
so it's readable when many tabs are open.

The page title should read more naturally (e.g. "package io").

Fixes b/145124550.

Change-Id: Iaffedd3b7c7f9944fb13c528e4229b2a29481b25
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/606104
Reviewed-by: Julie Qiu <julieqiu@google.com>