internal: add field sets
Add the ability to selectively read fields from the database.

Use a bitset (internal.FieldSet) to represent whether to read certain
potentionally large fields. The caller can choose whether to read
those fields from the DB by passing in a FieldSet to a database
method.

Represent an unread field in a struct with a special value. The
alternative of using a bitset in the struct to describe which fields
were absent is more cumbersome than we need. Currently the only fields
we care about are strings, so an unusual string value ("!MISSING") is
simpler, and will avoid serious bugs: at worst we'll render "!MISSING"
on the page, but the rest of the information will still be there.

Currently the only fields supported are ReadmeContents and
DocumentationHTML, and the only DB method with this ability is
GetDirectory.

In this CL, no use is made of this feature; we always read
all the fields.

Updates b/144281274.

Change-Id: I75bad07699d63c401899143b5fe69cd6ddea01a7
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/618420
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>