internal/etl: use multiple batches when writing completion indexes
Our AppEngine instances are limited to 2GiB of ram, and sometimes OOM
when writing completion data. This causes completions to be left in a
partial state, which then causes all subsequent updates to defensively
fail early.

The completion handler is updated to instead process completions in
multiple batches, reducing the total memory overhead of the operation
accordingly. The number of batches was chosen to reduce memory
consumption, but not have so many batches that the entire operation was
significantly slowed down.

Also check rows.Err before proceeding with copying over completions.

Updates b/146342339

Change-Id: I8c382d836f6f4eaa72c7c7603b97ea24d83544dd
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/627233
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>