migrations: add redistributable column to version
Add a column to the versions table to record whether a module is
redistributable.

(The packages table already has this column; it is populated but
unused.)

The check for whether a module is redistributable has become more
complex, possibly requiring arbitrary files in the zip. We can't and
shouldn't be doing any of it on the frontend.

Change-Id: Ia484891d4d48cef5ed8dbf2d8baa970ac8d8a3ce
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/628053
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>