internal/proxy,fetch: extract module path from go.mod more carefully
Parsing the go.mod file can fail, which caused us to give a 500 status
to the module.

We could just mark these with BadModule, but sometimes parsing fails
even for a syntactically valid go.mod file. For example, a file with the line

  require "gopkg.in/yaml.v2" v0.0.0-20180109114331-d670f9405373

will fail to parse with an error about how gopkg.in/yaml.v2 must have
a v2 version.

It seems too harsh to mark the module as bad just for that, so instead
of parsing we use modfile.ModulePath to extract the module path.  If
that fails we mark the module bad.

Change-Id: Ic379aaca8658ee2764f002aabbf33264b4fc5c9a
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/633302
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>