internal/licenses: reduce thresholds for Apache
Reduce the coverage and classification thresholds for Apache 2.0
licenses.

This works around the fact that the licensecheck package
does not have a template for the Apache 2.0 license without
the Appendix, so it has a reduced match score for it.
Both versions of Apache 2.0 are common in the wild.

Fixes b/147599731.

Change-Id: I4deb47748054f06bfe8984fa51e2b9973bc8d823
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/641360
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Hana Kim <hakim@google.com>