internal/fetch: update to go/doc from go1.14rc1
Start using the latest go/doc package as available in Go 1.14 RC 1.
It contains functionality to classify examples and accept slice of
parsed Go files without needing to go through an *ast.Package type.
That functionality was upstreamed from here and has gone through
Go standard library code review in https://golang.org/cl/204830.

It also includes changes from CLs https://golang.org/cl/211357 and
https://golang.org/cl/211598 to skip example functions with arguments.
Such functions are not valid examples and should not be included when
rendering Go package documentation.

One goal was to make this change earlier (before 1.14 is released) in
order to detect any issues with the API or functionality, and be able
to fix it in the standard library if needed. I have not discovered any
problems so far, and the tests added in the previous CL are passing.

Updates b/137567588.

Change-Id: I52e92f5e532ca91f39c6a5b40ecf501ac9a54a59
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/657662
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>