internal/licenses: support exception files
  Add a second kind of exception: file contents.

Instead of just matching a particular set of pathnames and their
contents tied to a particular module, also accept some license
contents wherever they appear.

Also, change the gonum license from the first kind to the second.

The result is that we recognize the gonum license as redistributable
anywhere it appears, not just at the root of the gonum module.

For example, now we will recognize it in
k8s.io/kubernetes/third_party/forked/gonum/graph, where before we
wouldn't because the exception only applied to the gonum.org/v1/gonum
module itself.

Change-Id: I554a2eca3df40f0d62ab5cf703fb8e8bbc5a6cb2
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/645960
Reviewed-by: Hana Kim <hakim@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>