internal/postgres: determine latest version correctly
The logic for determining whether the version being inserted is the
latest was incorrect: it would return true if the inserted version
were non-release and the existing latest version was release.

This caused the contents of imports_unique to reflect the wrong version
for such modules.

Change-Id: I4d69b9a36865896897e99a8d9ea70c17b2a7d9fb
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/655401
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>