internal/{etl,stdlib}: add test coverage for example classification
There wasn't much test coverage for example classification in the
existing test cases, making it so that subtle changes to that
functionality could get through undetected.

The encoding/json package has a good amount of examples, which makes
it a good test case for example classification and rendering.

Add specific test cases that include surrounding HTML in order to
increase test coverage, especially as the next CL will be making
a change to the go/doc package implementation (using the version
from Go 1.14 sooner). If these tests prove to be too rigid in the
future, they can be relaxed accordingly, but having more coverage
should help with making a safe go/doc transition.

Also fix a Markdown formatting typo in the README.

Updates b/137567588.

Change-Id: Icec2eb8d075db735f09967fb0c5e6991978ab0e2
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/657661
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>