internal/postgres: return all experiments, not just active experiments
Currently, in order to be able to enable an experiment using the query
string, the experiment must be considered active (meaning rolled out to
at least 1%). I think it makes more sense to just consider an experiment
active if it exists in the database. We can always delete experiment
rows if we no longer need them.

Change-Id: I1bc08ca11185b71bf370a0ec1cfbacff712365f1
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/658562
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>