internal/{etl,postgres}: insert package version state to postgres
The following methods are added to internal/postgres:
- GetPackageVersionStatesForModule
- GetPackageVersionState

UpsertModuleVersionState is modified, so that it will also
insert the provided package version states into the database.

The ETL is modified to use these methods to insert data into
package_version_states.

A bug is also fixed in the ETL, where the requestVersion was being
inserted into module_version_states and used to find a search_document
row, instead of the resolved version.

Updates b/135954292

Change-Id: I01e74ba26e95538069a4d34137f99a5e8be074dc
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/663069
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>