cmd/etl: insert a hard-coded list of exclusions
Permanent exclusions (user requests, malicious license violations, etc.) are
at risk of vanishing if the DB is corrupted or truncated. Keep the
list in code (it's pretty small), and insert it on etl startup.

We clumsily check if it's present, then insert if it isn't. The alternatives
are: (1) check the error to see if it is a primary key violation, but that
requires looking into the pq package's errors, and we don't want to do that
(what if we switch DB clients?); or (2) change InsertedExcludedPrefix to
do some other conflict resolution, but that wouldn't be right for its
primary use with dbadmin.

Change-Id: If346d97326b3ef566eb0c9ce9387c05645e47955
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/663154
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>