go.mod: update licensecheck
Update licensecheck to the latest version.

The new version of licensecheck is both more accurate and more flexible
than the old. It allows us to provide our own set of licenses, but
it is also more accurate about coverage.

I looked for changes this would cause in our production DB by
running

  go run ./devtools/cmd/license -db prod -lf /tmp/prod-licenses.gob -v changes






  


  


  
  






Change-Id: Ica256018f7b9e3846da63a94fb3d3b0ddf546fa7
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/665791
Reviewed-by: Julie Qiu <julieqiu@google.com>