internal/middleware: protect quota's cache with a mutex
The LRU cache was being accessed concurrently, but is not safe for
concurrent access. See

for the kind of crash that ensues.

Protect operations on it with a mutex. Locking the entire check-add
sequence ensures that we don't racingly create two limiters for the
same key.

Change-Id: I284439901eb061568f63d08b5af596ebc21c4560
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/669748
Reviewed-by: Julie Qiu <julieqiu@google.com>
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>