config,cmd/frontend: try a secondary DB host if the main one is down
- Provide a way to configure a secondary DB host.

- Modify the frontend to try connecting to the secondary if the primary fails.

This is a partial solution to the problem discussed in the bug below. It deals
with DBs that are bad on startup, but not ones that go bad later. For that case,
we have to wait for the process to die, or kill it manually, so on restart it
can pick the secondary DB.

Updates b/149906513.

Change-Id: I177590111dc4544476a4b2aef39f06456ba28ab5
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/673861
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>