internal/licenses: redo exceptions
Match our exception files ourselves, instead of adding them to
licensecheck.

Russ Cox suggested that we keep our own list of exception files and
match them ourselves, instead of adding them to the licensecheck
corpus. Our exceptions aren't really new license types, so shouldn't
be analyzed by licensecheck. And each file added to the corpus slows
down the checker.

Since we expect many exceptions, put the files in a directory
and process them with go generate. To avoid keeping the full
text of all the exceptions in memory as map keys, compute
the SHA of the normalized license text.

Change-Id: Ia8464968aac5a43fd27a8324cb37e3ad4b0f3e99
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/677789
Reviewed-by: Robert Findley <rfindley@google.com>