internal/frontend: handlers return error
Instead of serving an error page themselves, top-level handlers return
an error with the information for serving the error page.

This is a largely mechanical rewrite that removes a lot of clumsiness
from the handling code. For example, instead of

   if err != nil {
      log.Errorf(ctx, "frobbing: %v", err)
      s.serveErrorPage(w, r, http.StatusInternalServerError, nil)
      return
   }

the code is both simpler and more idiomatic:

   if err != nil {
       return fmt.Errorf("frobbing: %v", err)
   }

It is still possible to include an errorPage with the error, like so:

  if err != nil {
      return &serverError{
          status: http.StatusBadRequest,
	  epage: &errorPage{...},
	  err: err,
      }
  }

Change-Id: Id076894b1cb912fe0731fdf6fbdb7b0e5b972bcb
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/686039
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>