migration: add time zone to timestamp columns (smaller tables)
The "timestamp without time zone" type doesn't properly identify
a point in time. You need the time zone for that.

This migration updates some of our "timestamp without time zone" columns
to "timestamp with time zone".

We know that in fact, the timestamps are stored in UTC, so that is the
time zone we use in the conversion.

This migration updates only the relatively small tables, because migrating
all the tables takes 3 hours.

We'll do the remaining tables each in their own migration.

See https://wiki.postgresql.org/wiki/Don%27t_Do_This#Don.27t_use_timestamp_.28without_time_zone.29.

Change-Id: Ifb05c413b577b0bf3c46a6c3f4aaade8b7f8a7e7
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/692962
Reviewed-by: Julie Qiu <julieqiu@google.com>