internal/fetch/dochtml/internal/render: insert '¶' permalinks in headings
Make it an option that can be toggled, so it is easier to tell that
this feature is different from the original CL 72890. This option
can be deleted later when there are fewer godoc variants and the
flexibility it offers becomes no longer needed.

This package is internal, so changing its API can be done at any time.

Background

The render package was initially developed in the context of the
golang.org/x/tools/cmd/godoc command. That command historically
used the ToHTML function from go/doc package, which did not have
functionality to insert permalinks for headings. That functionality
was later implemented in https://golang.org/cl/33641 on the frontend
using JavaScript.

It was done on the frontend because modifying the go/doc package was
slow, and because the intent was to make a change affecting the entire
golang.org website.

Here, we have full control over the documentation HTML rendering, so
it's easy to implement this functionality on the backend for the needs
of the discovery site.

Fixes golang/go#36808.

Change-Id: Icb0210f9e0870a7bbf80e0aa80ba2e6421177ba8
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/698511
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>