migrations: set B weight to 1
Weight the B section of the search document the same as the A section.

This will the synopsis and the path equal weight, which will improve
search results.

This CL is just preparation for that. It won't take effect until
we change the code to call the popular_search function instead of
popular_search_go_mod.

Change-Id: Ic5ca03ef1296a4852c80d9aa1c145644acd7dd9c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/705158
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>