internal/proxy: delete license testadata
proxy.TestModule now uses license testdata that already lives in
internal/testing/testdata.

Any TestModule that previously had a BSD3
license now has a BSD0 license, since the BSD0 license was already
present in internal/testing/testdata, and the type of
license for those tests did not matter, as long as it was redistributable.

Change-Id: Ic8837b2fe54e51e52086c859b1937de3269f8d2b
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/710138
Reviewed-by: Jonathan Amsterdam <jba@google.com>