internal/middleware: set SameSite on temp godoc cookie to Lax
If SameSite is set to Strict on a cookie, then it won’t be sent by the
browser if the request is initiated via a link from another site. This
scenario plays out when a user originates from godoc.org and
eventually lands on the right pkg.go.dev page via a chain of
redirects. As a result, the temporary cookie is never sent to the
server when it needs to be and the “Back to godoc.org” button is
not shown.

Update the SameSite attribute to Lax to allow the desired behavior.

Also, update the redirect method becuase 302 is more appropriate
for this use-case (we’re not redirecting POST requests).

More information about the SameSite attribute is at
https://web.dev/samesite-cookies-explained/

Updates b/147492306

Change-Id: If344dcca1191b08c9a536d98837d45561cc08bba
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/710838
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>