migrations: remove golang text configuration
The golang text configuration was intended to prevent stemming for
terms like "postgres" and "NATS". However, it's too great a loss to
drop stemming for other, ordinary English words (e.g. "logging" to
"log"). And though "postgres" is stemmed to the nonsensical "postgr",
that happens for both the document and the search query, so search
quality isn't affected.

Change-Id: I730590e27499fca9f70a3ce65a6e6ea364f4adda
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/712880
Reviewed-by: Julie Qiu <julieqiu@google.com>