migrations: add a text config for path tokens
Add a text configuration that ignores parts of hyphenated words, to
reduce the inflated search ranking of paths with hyphenated elements.

We'll use this to create part of a search document that is taken from
the package path.

Currently we split a path into tokens, generating all sub-paths as
well as splitting at hyphens. For example,

    github.com/CrunchyData/postgres-operator/apiserver

results in the tokens

    postgres
    postgres-operator
    postgres-operator/apiserver

among others. The default text configuration splits at hyphens too, so
the three tokens above result in three copies of "postgres", when we'd
expect only one. That skews the search results to favor this package
more than it should.

If we use a text configuration that ignores hyphenated word parts, the
above tokens will result in only once instance of "postgres".

Change-Id: Ide76d80d32c079b6e3a07c7434ddddd61b39aadf
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/712881
Reviewed-by: Julie Qiu <julieqiu@google.com>