internal/postgres: insert module_id and refactor saveModule
licenses.module_id and version_map.module_id are now populated.

saveModule is refactored so that inserting the module is done in
multiple functions, instead of one large function, for readability. The
functionality has not change and the pieces of a module are still being
inserted in a single transaction.

Change-Id: I3603a388be5dbb90ce8f05ae9c237989f965e6e4
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/717237
Reviewed-by: Jonathan Amsterdam <jba@google.com>